增加部分注释信息

+++ Classes/Base/POPTStringUtils.cpp
@@ -12,8 +12,9 @@ vector<string> POPTStringUtils::split(string str, string pattern){
     string::size_type pos;
     vector<string> result;
     str+=pattern;//扩展字符串以方便操作
-    int size=str.size();
+    int size=(int)str.size();
     
+
     for(int i=0; i<size; i++)
     {
         pos=str.find(pattern,i);
@@ -21,7 +22,7 @@ vector<string> POPTStringUtils::split(string str, string pattern){
         {
             string s=str.substr(i,pos-i);
             result.push_back(s);
-            i=pos+pattern.size()-1;
+            i=(int)(pos+pattern.size()-1);
         }
     }
     return result;
